fix compile error for lazy example
